Removed client.cpp anda moved to RandomHandler. Chnaged Bankaccounts members to float.

+++ src/BankAccount.cpp
@@ -1,9 +1,9 @@
 #include "../include/BankAccount.hpp"
 
 BankAccount::BankAccount() : balance(0.0), accountNumber(0) {}
-BankAccount::BankAccount(int accountNumber, double balance) : accountNumber(accountNumber), balance(balance) {}
+BankAccount::BankAccount(int accountNumber, float balance) : accountNumber(accountNumber), balance(balance) {}
 
-void BankAccount::deposit(double amount)
+void BankAccount::deposit(float amount)
 {
     if (amount > 0)
     {
@@ -11,7 +11,7 @@ void BankAccount::deposit(double amount)
     }
 }
 
-void BankAccount::withdraw(double amount)
+void BankAccount::withdraw(float amount)
 {
     if (amount > 0 && amount <= balance)
     {
@@ -19,7 +19,7 @@ void BankAccount::withdraw(double amount)
     }
 }
 
-double BankAccount::getBalance() const
+float BankAccount::getBalance() const
 {
     return balance;
 }
